% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxfit.R
\docType{package}
\name{coxfit-package}
\alias{coxfit}
\alias{coxfit-package}
\title{coxfit: Fit Cox Proportional Hazards Model for Ordered Survival Data}
\description{
Assuming that the data is sorted in order of increasing survival time, coxfit fits a Cox Proportional Hazards 	model given a design matrix and a censoring vector. It gives equal results to the coxph function of the survival 	package.
}
\details{
The only function you will actually run is fit_cox.
Please refer to the cox_fit documentation for more details and a short example.
}
\references{
\itemize{
\item coxfit: Fitting Cox proportional hazards model for survival data
\item \url{https://github.com/lamttran/coxfit}
}
}
\author{
\itemize{
\item Lam Tran <lamtran@umich.edu>
}
}
