% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cox.R
\name{fit_cox}
\alias{fit_cox}
\title{Fit a Cox proportional hazards model to survival data}
\usage{
fit_cox(x, censor)
}
\arguments{
\item{x}{Design matrix of data, assumes survival times sorted in INCREASING order}

\item{censor}{Censoring vector of data}
}
\value{
Vector of estimated parameters for each covariate
}
\description{
Fit a Cox proportional hazards model to survival data
}
\examples{
p = 5
n = 100
c0 = 0.5
beta = rnorm(p, 0, 1)
x = matrix(rnorm(n * length(beta)), ncol = length(beta))
yvar= rnorm(n, 0, 1)
times=rexp(n, c0 * exp(rowSums(t(t(x) * beta)) + yvar))
time.censor=rexp(n, c0 * exp(beta[1] * x[, 1] + yvar))
censorv=ifelse(times < time.censor, 1, 0)
y = ifelse(times < time.censor, times, time.censor)
ord = order(y, decreasing = FALSE)
dataset = list(x = x[ord,], y = y[ord], censor = censorv[ord])
fit_cox(dataset$x, dataset$censor)
all.equal(fit_cox(dataset$x, dataset$censor), as.numeric(coxph(Surv(dataset$y, dataset$censor) ~ dataset$x)$coefficients))

}
\keyword{Cox}
\keyword{hazards}
\keyword{proportional}
\keyword{survival,}
